+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -4,36 +4,40 @@ about: Suggest an idea for this project
 title: ''
 labels: ''
 assignees: ''
-
 ---
 
 ## **✨ Feature Request: [기능 이름] 추가**
 
 ### **📌 개요**
+
 현재 **[기존 상태]** 에서 **[새로운 기능이 필요한 이유]**  
 이를 해결하기 위해 **[새로운 기능]** 을 추가하려 합니다.
 
 ---
 
 ### **🎯 요구사항**
+
 ✅ **[기능 A]** → [구체적인 동작 방식]  
 ✅ **[기능 B]** → [구체적인 동작 방식]  
-✅ **[기능 C]** → [구체적인 동작 방식]  
+✅ **[기능 C]** → [구체적인 동작 방식]
 
 🔹 **예상되는 사용자 인터랙션**:
-1. [사용자가 어떤 행동]을 하면,  
+
+1. [사용자가 어떤 행동]을 하면,
 2. [어떤 변화]가 UI에서 발생해야 한다.
 
 ---
 
 ### **💡 기대 효과**
+
 ✔ **사용자 경험(UX) 개선**: [기능이 추가되면 UX가 어떻게 개선되는지]  
 ✔ **퍼포먼스 최적화**: [새로운 기능이 성능 개선에 어떤 영향을 미치는지]  
-✔ **비즈니스 가치 향상**: [이 기능이 서비스에 어떤 가치를 추가하는지]  
+✔ **비즈니스 가치 향상**: [이 기능이 서비스에 어떤 가치를 추가하는지]
 
 ---
 
 ### **🛠 작업 내용**
+
 - [ ] **UI 디자인 적용**
 - [ ] **기능 구현**
 - [ ] **API 연동**
@@ -42,6 +46,7 @@ assignees: ''
 ---
 
 ### **📌 참고 자료**
+
 📌 **디자인 링크**: [Figma, Zeplin 등]  
 📌 **API 문서**: [Swagger, Notion 문서 등]  
 📌 **관련 기술 스택**: [사용할 라이브러리, 기술 등]
